refactor: move stor values to poitions content file
